refactor: break template tests out into one file per template
Split the monolithic test/templates_test.rb into individual files
under test/templates/, one per template. Shared setup/teardown and
helpers (create_rails_app, apply_template, assert_rails_boots) move
to a TemplateTestCase base class in test_helper.rb. The Rakefile
glob widens to test/**/*_test.rb to pick up the subdirectory.

No test logic changes: 13 runs, 207 assertions before and after.
