+++ Rakefile
@@ -2,7 +2,7 @@ require "rake/testtask"
 
 Rake::TestTask.new(:test) do |t|
   t.libs << "test"
-  t.test_files = FileList["test/*_test.rb"]
+  t.test_files = FileList["test/**/*_test.rb"]
   t.verbose = true
 end
